crazy type warnings differ between linux and osx gcc, heh

+++ project/src/audio/lumen_audio_files.cpp
@@ -125,7 +125,7 @@ namespace lumen {
                 printf("bitrate lower   %lu \n",    ogg_source->info->bitrate_lower);
                 printf("bitrate window  %lu \n",    ogg_source->info->bitrate_window);
                 printf("vendor          %s \n",     ogg_source->comments->vendor);
-                printf("length          %ld \n",    ogg_source->length);
+                printf("length          %lld \n",    (long long)ogg_source->length);
                 printf("uncompressed    %lld \n",   (long long)total_length);
                 
             for(int i = 0; i < ogg_source->comments->comments; i++) {
